pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantcli"
version = "0.1.1"
description = "面向AI的多因子量化选股策略挖掘工具，AI Agent 友好 CLI"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "QuantCLI Team", email = "quantcli@example.com"}
]
urls = {repository = "https://gitcode.com/datavoid/quantcli"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=12.0",
    "backtrader>=1.9",
    "tabulate>=0.9",
    "pyyaml>=6.0",
    "pymysql>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
quantcli = "quantcli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["quantcli*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
